+++ web/src/App.tsx
@@ -12,7 +12,7 @@ import Loading from './pages/Loading';
 import AquaForms from './pages/AquaForms';
 import FormGenerator from './pages/FormGenerator';
 import AttestationAddresses from './pages/AttestationAddresses';
-import PdfSigner from './pages/PdfSigner/PdfSigner';
+import PdfSigner from './pages/wokflow/ContractDocument/PdfSigner';
 import WorkFlowPage from './pages/wokflow/WorkFlow';
 // import Aquafier from 'aqua-js-sdk';
 declare global {
@@ -46,7 +46,7 @@ function App() {
           <Route path="/loading" element={<Loading />} />
           <Route path="/share/:identifier" element={<SharePage />} />
           <Route path="/aqua-forms" element={<AquaForms /> } />
-          <Route path="/pdf-signer" element={<PdfSigner  file={null}  submittingSignatureData={false} submitSignature={async(_one, _two)=>{
+          <Route path="/pdf-signer" element={<PdfSigner  file={null}  setActiveStep={(_one)=>{
 
           }}/> } />
           <Route path="/workflow" element={<WorkFlowPage /> } />
